backend/requirements.txt
django~=3.2.13
djangorestframework~=3.13.1
django-filter~=21.1
django-redis~=5.2.0
celery~=5.2.6
channels~=3.0.5
channels-redis~=3.4.0
daphne~=3.0.2
drf-yasg~=1.20.0
websocket-client~=1.3.3
psycopg2~=2.9.3
requests~=2.28.1